project_podcast_summarizer/requirements-local.txt
llama-cpp-python>=0.2.44,<0.3.0
llama-index==0.9.39
transformers==4.37.1
deepeval==0.20.55
pytest==8.0.0
typer>=0.9.0,<1.0.0
